wally.repl
Utilities for interacting with a page from the REPL while in the middle of a test (or other script).

Found it useful when developing tests to be able to pause inside a test with all the context, fixtures etc. in place.

This is very generic stuff and maybe doesn't belong to this library... OTH, maybe it does as it is probably a very common use case.