Add `wait-for-not-visible`
This is just a wrapper around `wait-for`, but makes it more
obvious how to wait for things to *not* be on the page.

The idea to return `true` comes from the fact that obviously
in case of success `wait-for` would return `nil` as we don't
expect to find anything. Now if we wanted to have this call
in `is` in a test it's easier to just return `true` here.