Allow subqueries in `query`
If we pass to `query` a vector starting with a locator
or `SequableLocator`, the remaining part of the vector
is used as a subquery in the subtree(s) rather than
the entire page.

An example included at the end of the file.

While there, I've replaced the check for concrete
`LocatorImpl` with a check for its interface.